python/loglog/requirements.txt
loguru==0.7.3
pyzmq==26.2.0
aiohttp==3.11.11
aiorun==2024.8.1
elasticsearch==7.16.3
python-dotenv==1.0.1
tornado==6.4.2